Moved embed finders into separate modules
