+++ wagtail/wagtailembeds/exceptions.py
@@ -1,0 +1,6 @@
+class EmbedException(Exception):
+    pass
+
+
+class EmbedNotFoundException(EmbedException):
+    pass
