+++ wagtail/wagtailembeds/templatetags/wagtailembeds_tags.py
@@ -2,7 +2,7 @@ from django import template
 from django.utils.safestring import mark_safe
 
 from wagtail.wagtailembeds import embeds
-
+from wagtail.wagtailembeds.exceptions import EmbedException
 
 register = template.Library()
 
@@ -12,5 +12,5 @@ def embed(url, max_width=None):
     try:
         embed = embeds.get_embed(url, max_width=max_width)
         return mark_safe(embed.html)
-    except embeds.EmbedException:
+    except EmbedException:
         return ''
